feat(voice-ai): Integrate ElevenLabs for text-to-speech
Implement ElevenLabs API for enhanced text-to-speech within the `VoiceAI` class.
Adds ElevenLabs API key and voice ID parameters to enable external TTS.
Includes a fallback to native browser `SpeechSynthesis` if ElevenLabs is not configured or an error occurs.
Manages audio playback via `AudioContext` and updates `disconnect` logic for ElevenLabs audio sources.
Introduces `.env` for API key management.
